Updated this project to the new standard, see details.
After a significant modification to the Codebase library, I imported a lot
of the Codebase updates back to this porject. Some files that exist in
Codebase still carry the zlib license header; these files are still under
the zlib license.

Following a significant struggle with SDL_net, I've decided to ditch that
library and write my own version. It currently only supports TCP
networking, and only works on windows platforms,  but it is sufficient at
this time.