Initial commit, read more
After a few days of work, I'm committing what i've done so far. Mostly
I've cherry picked the best parts of the multiplayer branch's modules, and
adapted them for use here. I've written a few new components, and I've
learnt a few things along the way. I may continue with this branch for a
while, before moving the changes into the main branch. Alternatively, it
might turn out that this single player version is more fun. We'll see.

I've swapped the config system for a pure lua version instead. This is
actually fairly flexible. I've also changed the map system and it's API a
great deal. The tile sheet can be loaded from lua, and the save directory
isn't hard coded anymore. I think this would be a good testbed for a lot
of different systems, including the collision system, however the current
system (using vectors) is inadequate. I'll have to dig up the old BBox,
possilby.

This branch will piggy back on the main repo for a while.